requirements.txt
Flask
notion-client
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openai
requests
discord.py
gunicorn
